Added discovery function and icons. Also added listeners to control your tv with a keyboard. (#1)
* Added discovery function and icons.

Still need to find a power icon and get more of the icons from the original remote

* Improved icons and added new device functionality

* added roku image back in

* cleanup. Added listener for enter

* Fixed index issue and added "Add TV" button to dropdown

* embed image to keep app portable

* Fixed add tv bug. Added new IP func, and added cleaned up button command timeout

The default command timeout made it seem like the application was freezing. Reduced that to 1 second which is still a long time, but works well. Also added a dialog for that same thing.

* updated embedded image section

* Added discovery context(device names)